requirements.txt
certifi==2020.6.20
chardet==3.0.4
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
idna==2.10
isort==5.9.3
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
pyTelegramBotAPI==3.7.3
python-dotenv==0.19.0
requests==2.24.0
six==1.15.0
testfixtures==6.18.1
urllib3==1.25.10
